requirements.txt
arrow
click
dictdiffer
pymongo
xxhash
pyyaml
git+https://github.com/jmcgeheeiv/pyfakefs.git